Remove ALIGN_LONG and simplify code
So turns out that ARM aligned loads are quite fun. Running Genesis
(built for armv7) on an armv8 kernel, results in crashes due to stm/ldm
instructions. The reason is that armv7 kernels "fixup" the unaligned
instructions but armv8 kernels don't do that anymore.

This patch falls back to code that always work in all platforms. In fact
modern gcc/clang produces the same optimized code as before (a single
load/store) in platforms that support it like x86 or powerpc.