[Libretro] Rework "Hide Master System Left Border" option
Reworked the option so that now you can hide either 8 pixels on both left and right sides of the screen, or the 8 pixels on the left side only.
Also added a check that makes sure it doesn't cut whenever VDP register $00 bit 5 isn't set, which prevents the option from cutting when it's not necessary (e.g. the title screen of Gangster Town, the half pipe stage of California Games, the Phantasy Star menu screen).